fix(WaterfallChart): stop casting the y column to float in standalone code
`createPlotlyFigure` reads the y column as it comes; the standalone path cast
it with `.astype(float)`. The cast is not a harmless widening: the figure
labels each bar with `f"{v:+}"`, so an INTEGER y renders "+1" on the platform
and "+1.0" standalone.

`yColumn` accepts integer, long and double, and the auto tier resolves it to
`id` against the canonical fixture, so the divergent case is the one the
verification run exercises. Without the cast the operator verifies clean.

+++ common/workflow-operator/src/main/scala/org/apache/texera/amber/operator/visualization/waterfallChart/WaterfallChartOpDesc.scala
@@ -134,7 +134,7 @@ class WaterfallChartOpDesc extends PythonOperatorDescriptor with StandaloneCodeG
        |else:
        |    table = in1df
        |    x_values = table["$xColumn"]
-       |    y_values = table["$yColumn"].astype(float)
+       |    y_values = table["$yColumn"]
        |
        |    fig = go.Figure(go.Waterfall(
        |        name="Waterfall", orientation="v",
